Add AlecoAir PU55 HUMINO device configuration
Added configuration for AlecoAir PU55 HUMINO Purifier-Humidifier including various entities and their properties.

PU55 Humino is a 2 in 1 Purifier Humidifier. The humidifier function can be engaged only if the purifier is powered on . Also it has 2 humidifier functions. 1st is a constant mist ( Humidify ) and the second is based on the current humidity ( Humidify Controlled ) 

Setting dps: 105 to on will activate the controlled mode and also dps 10 will become on automatically
Setting dps: 10 to on will start a permanent mist